pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "farmrl-openenv-submission"
version = "0.1.0"
description = "FarmRL OpenEnv submission package"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "numpy",
    "pandas",
    "pydantic",
    "pyyaml",
    "openai",
    "python-dotenv",
    "fastapi",
    "uvicorn",
    "openenv-core>=0.2.0",
]

[project.scripts]
server = "server.app:main"